markdown-hide-body no longer alias for outline-...
Either I switched from a hacked version of jblevin's mode to the
original or the alias for markdown-hide-body to outline-hide-body has
gone away.

I've switched the setting in the file local variable in words.gpg to
use outline-hide-body and updated my Emacs config to allow it to be
used as a file-local value.